--- kata/ditto-java-client/policy-management/src/main/java/com/bosch/iot/things/kata/package-info.java
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) 2020 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-
-/**
- * The content of this package is of general purpose and thus subject to be moved in its own library in future
- */
-package com.bosch.iot.things.kata;
